Fix for session participants
The internal SessionParticipants is not exposed, so we should assume they are sending us an array of session participants rather than an instance of the SessionParticipants. If we send an array, currently we get an error similar to "options.sessionParticipants.get is not a function" because it's an array instead of the unexposed SessionParticipants.