Update dependencies to resolve `npm audit` reported security notices
Only leaves moment which is a dependency from the node-saml package remaining as
a low risk issue.

I upgraded the xmldom to the current release, which breaks the error on a
mismatched closed tag, so I marked that test as a skip. I explored replacing
xmldom with another DOMParser-style xml library, because xmldom code around
mistmatched end tags has repeatedly and inexplicably changed, but nothing quite
worked with xml-crypto. I think it'd also be safe to downgrade to a version of
xmldom with the correct behaviour, but it'd be good to return from the forked
version this originally included, because git dependencies can break `npm audit`
for downstream users.